перенос всех типов в index
