+++ src/pages/indredient-card.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useAppSelector } from '../hooks/redux';
 import { useParams } from 'react-router-dom';
 import IngredientDetails from '../components/ingredient-details/ingredient-details'
-import { IIngredient } from '../types/request-types';
+import { IIngredient } from '../types/index';
 
 const IndredientCard:React.FC = () => {
