+++ src/services/actions/ingredients.ts
@@ -1,7 +1,7 @@
 import { v4 as uuid } from 'uuid'
 import { checkResponce } from '../../utils/check-responce';
-import { IIngredient, IIngredientUUID } from '../../types/request-types';
-import {  } from '../../types/request-types';
+import { IIngredient, IIngredientUUID } from '../../types/index';
+import {  } from '../../types/index';
 import { getIngredientRequest } from '../api';
 import { AppDispatch } from '../store';
