+++ src/types/index.ts
@@ -37,3 +37,33 @@ export interface IIngredient{
     image_large: string,
     __v: number,
 }
+
+export interface IUser{
+    email: string,
+    name: string,
+}
+
+export interface IIngredientUUID extends IIngredient{
+    uuid: string
+}
+
+export interface IOrder{
+    success: boolean;
+    name: string;
+    order: {
+        ingredients: Array<IIngredient>;
+        _id: string;
+        owner: {
+            name: string;
+            email: string;
+            createdAt: string;
+            updatedAt: string;
+        }
+        status: 'done' | 'pending' | 'created';
+        name: string;
+        createdAt: string;
+        updatedAt: string;
+        number: number;
+        price: number;
+    }
+}
